+++ app/Http/Controllers/Manager/StatisticsGlobalController.php
@@ -1,7 +1,7 @@
 <?php
 
 /**
- * Copyright (c) 2018-2021 Adshares sp. z o.o.
+ * Copyright (c) 2018-2024 Adshares sp. z o.o.
  *
  * This file is part of AdServer
  *
@@ -24,42 +24,51 @@ declare(strict_types=1);
 namespace Adshares\Adserver\Http\Controllers\Manager;
 
 use Adshares\Adserver\Http\Controller;
+use Adshares\Adserver\Models\TurnoverEntry;
 use Adshares\Adserver\Repository\Common\TotalFeeReader;
 use Adshares\Adserver\Repository\Demand\MySqlDemandServerStatisticsRepository;
 use Adshares\Adserver\Repository\Supply\MySqlSupplyServerStatisticsRepository;
 use Adshares\Common\Exception\InvalidArgumentException;
+use Adshares\Supply\Domain\ValueObject\TurnoverEntryType;
+use DateTimeImmutable;
+use DateTimeInterface;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Storage;
 use Symfony\Component\HttpFoundation\StreamedResponse;
 use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
+use Symfony\Component\HttpKernel\Exception\UnprocessableEntityHttpException;
 
 class StatisticsGlobalController extends Controller
 {
-    /** @var MySqlDemandServerStatisticsRepository */
-    private $demandRepository;
-
-    /** @var MySqlSupplyServerStatisticsRepository */
-    private $supplyRepository;
-
-    /** @var TotalFeeReader */
-    private $totalFeeReader;
+    private const DEMAND_EXPENSE_TYPES = [
+        TurnoverEntryType::DspExpense,
+        TurnoverEntryType::DspJoiningFeeAllocation,
+    ];
+    private const SUPPLY_EXPENSE_TYPES = [
+        TurnoverEntryType::SspJoiningFeeExpense,
+    ];
+    private const SUPPLY_OPERATOR_INCOME_TYPES = [
+        TurnoverEntryType::SspBoostOperatorIncome,
+        TurnoverEntryType::SspOperatorFee,
+    ];
+    private const SUPPLY_PUBLISHERS_INCOME_TYPES = [
+        TurnoverEntryType::SspBoostLocked,
+        TurnoverEntryType::SspPublishersIncome,
+    ];
 
     public function __construct(
-        MySqlDemandServerStatisticsRepository $demandRepository,
-        MySqlSupplyServerStatisticsRepository $supplyRepository,
-        TotalFeeReader $totalFeeReader
+        private readonly MySqlDemandServerStatisticsRepository $demandRepository,
+        private readonly MySqlSupplyServerStatisticsRepository $supplyRepository,
+        private readonly TotalFeeReader $totalFeeReader,
     ) {
-        $this->demandRepository = $demandRepository;
-        $this->supplyRepository = $supplyRepository;
-        $this->totalFeeReader = $totalFeeReader;
     }
 
-    public function fetchDemandStatistics()
+    public function fetchDemandStatistics(): array
     {
         return $this->demandRepository->fetchStatistics();
     }
 
-    public function fetchDemandDomains(Request $request)
+    public function fetchDemandDomains(Request $request): array
     {
         $days = max(1, min(30, (int)$request->get('days', 30)));
         $offset = max(0, min(30 - $days, (int)$request->get('offset', 0)));
@@ -67,24 +76,43 @@ class StatisticsGlobalController extends Controller
         return $this->demandRepository->fetchDomains($days, $offset);
     }
 
-    public function fetchDemandCampaigns()
+    public function fetchDemandCampaigns(): array
     {
         return $this->demandRepository->fetchCampaigns();
     }
 
-    public function fetchDemandBannersSizes()
+    public function fetchDemandBannersSizes(): array
     {
         return $this->demandRepository->fetchBannersSizes();
     }
 
-    public function fetchSupplyStatistics()
+    public function fetchDemandTurnover(string $from, string $to): array
+    {
+        $from = DateTimeImmutable::createFromFormat(DateTimeInterface::ATOM, $from);
+        $to = DateTimeImmutable::createFromFormat(DateTimeInterface::ATOM, $to);
+        if (false === $from || false === $to) {
+            throw new UnprocessableEntityHttpException('Invalid date format');
+        }
+
+        $turnoverEntries = TurnoverEntry::fetchByHourTimestamp($from, $to);
+        $expense = 0;
+        foreach ($turnoverEntries as $entry) {
+            if (in_array($entry->type, self::DEMAND_EXPENSE_TYPES, true)) {
+                $expense += $entry->amount;
+            }
+        }
+
+        return compact('expense');
+    }
+
+    public function fetchSupplyStatistics(): array
     {
         $totalFee = $this->totalFeeReader->getTotalFeeSupply();
 
         return $this->supplyRepository->fetchStatistics($totalFee);
     }
 
-    public function fetchSupplyDomains(Request $request)
+    public function fetchSupplyDomains(Request $request): array
     {
         $days = max(1, min(30, (int)$request->get('days', 30)));
         $offset = max(0, min(30 - $days, (int)$request->get('offset', 0)));
@@ -94,7 +122,34 @@ class StatisticsGlobalController extends Controller
         return $this->supplyRepository->fetchDomains($totalFee, $days, $offset);
     }
 
-    public function fetchSupplyZonesSizes()
+    public function fetchSupplyTurnover(string $from, string $to): array
+    {
+        $from = DateTimeImmutable::createFromFormat(DateTimeInterface::ATOM, $from);
+        $to = DateTimeImmutable::createFromFormat(DateTimeInterface::ATOM, $to);
+        if (false === $from || false === $to) {
+            throw new UnprocessableEntityHttpException('Invalid date format');
+        }
+
+        $turnoverEntries = TurnoverEntry::fetchByHourTimestamp($from, $to);
+        $expense = 0;
+        $operatorIncome = 0;
+        $publishersIncome = 0;
+        foreach ($turnoverEntries as $entry) {
+            if (in_array($entry->type, self::SUPPLY_EXPENSE_TYPES, true)) {
+                $expense += $entry->amount;
+            }
+            if (in_array($entry->type, self::SUPPLY_OPERATOR_INCOME_TYPES, true)) {
+                $operatorIncome += $entry->amount;
+            }
+            if (in_array($entry->type, self::SUPPLY_PUBLISHERS_INCOME_TYPES, true)) {
+                $publishersIncome += $entry->amount;
+            }
+        }
+
+        return compact('expense', 'operatorIncome', 'publishersIncome');
+    }
+
+    public function fetchSupplyZonesSizes(): array
     {
         return $this->supplyRepository->fetchZonesSizes();
     }