+++ tests/app/Models/SspHostTest.php
@@ -23,8 +23,10 @@ declare(strict_types=1);
 
 namespace Adshares\Adserver\Tests\Models;
 
+use Adshares\Adserver\Models\Config;
 use Adshares\Adserver\Models\SspHost;
 use Adshares\Adserver\Tests\TestCase;
+use Adshares\Adserver\Utilities\DatabaseConfigReader;
 
 class SspHostTest extends TestCase
 {
@@ -37,5 +39,27 @@ class SspHostTest extends TestCase
         $sspHost->ban();
 
         self::assertTrue($sspHost->banned);
+    }
+
+    public function testFetchAccepted(): void
+    {
+        SspHost::create('0001-00000001-8B4E', true);
+        SspHost::create('0001-00000002-BB2D', true);
+
+        $accepted = SspHost::fetchAccepted(config('app.inventory_export_whitelist'));
+
+        self::assertCount(2, $accepted);
+        $addresses = $accepted->map(fn($sspHost) => $sspHost->ads_address)->toArray();
+        self::assertContains('0001-00000001-8B4E', $addresses);
+        self::assertContains('0001-00000002-BB2D', $addresses);
+
+        Config::updateAdminSettings([Config::INVENTORY_EXPORT_WHITELIST => '0001-00000001-8B4E']);
+        DatabaseConfigReader::overwriteAdministrationConfig();
+
+        $accepted = SspHost::fetchAccepted(config('app.inventory_export_whitelist'));
+
+        self::assertCount(1, $accepted);
+        $addresses = $accepted->map(fn($sspHost) => $sspHost->ads_address)->toArray();
+        self::assertContains('0001-00000001-8B4E', $addresses);
     }
 }